backend/requirements.txt
flask
flask-cors
sqlalchemy
flask-sqlalchemy
psycopg2-binary
python-dotenv
requests